Move `local_global_conversion` (#213)
* Move local_global_conversion.py and its dependency to common repo
* Refactored imports
