+++ modules/flight_interface/flight_interface.py
@@ -2,13 +2,13 @@
 Creates flight controller and combines odometry data and timestamp.
 """
 
-from . import local_global_conversion
-from .. import drone_odometry_local
-from .. import odometry_and_time
 from .. import decision_command
+from .. import odometry_and_time
 from ..common.logger.modules import logger
 from ..common.mavlink.modules import drone_odometry
+from ..common.mavlink.modules import drone_odometry_local
 from ..common.mavlink.modules import flight_controller
+from ..common.mavlink.modules import local_global_conversion
 
 
 class FlightInterface:
