+++ modules/odometry_and_time.py
@@ -4,7 +4,7 @@ Drone odometry in local space and timestamp.
 
 import time
 
-from . import drone_odometry_local
+from .common.mavlink.modules import drone_odometry_local
 
 
 class OdometryAndTime:
