+++ tests/integration/test_data_merge_worker.py
@@ -7,14 +7,13 @@ import time
 
 import numpy as np
 
-from modules import drone_odometry_local
 from modules import detections_and_time
 from modules import merged_odometry_detections
 from modules import odometry_and_time
+from modules.common.mavlink.modules import drone_odometry_local
 from modules.data_merge import data_merge_worker
 from utilities.workers import queue_proxy_wrapper
 from utilities.workers import worker_controller
-
 
 DATA_MERGE_WORKER_TIMEOUT = 10.0  # seconds
