+++ tests/integration/test_geolocation_worker.py
@@ -8,13 +8,12 @@ import time
 import numpy as np
 
 from modules import detections_and_time
-from modules import drone_odometry_local
 from modules import merged_odometry_detections
+from modules.common.mavlink.modules import drone_odometry_local
 from modules.geolocation import camera_properties
 from modules.geolocation import geolocation_worker
 from utilities.workers import queue_proxy_wrapper
 from utilities.workers import worker_controller
-
 
 WORK_COUNT = 3
