+++ tests/unit/test_geolocation.py
@@ -7,12 +7,11 @@ import pytest
 
 from modules import detection_in_world
 from modules import detections_and_time
-from modules import drone_odometry_local
 from modules import merged_odometry_detections
+from modules.common.logger.modules import logger
+from modules.common.mavlink.modules import drone_odometry_local
 from modules.geolocation import camera_properties
 from modules.geolocation import geolocation
-from modules.common.logger.modules import logger
-
 
 FLOAT_PRECISION_TOLERANCE = 4
